Mostly cosmetic changes that enable plplot to compiled with (almost) no warnings, even with gcc -Wall.
Most changes are just casts, and most of them are tcl/tk related. For tcl/tk-8.4, no warnings occurs.
Also tested with tcl/tk-8.3, where some warnings remain.
There are no java/f77/cxx/python/octave changes.

svn path=/trunk/; revision=4991

+++ src/pdfutils.c
@@ -437,7 +437,7 @@ pdf_wr_string(PDFstrm *pdfs, const char *string)
 
     dbug_enter("pdf_wr_string");
 
-    for (i = 0; i <= strlen(string); i++) {
+    for (i = 0; i <= (int) strlen(string); i++) {
 	if (pdf_putc(string[i], pdfs) == EOF)
 	    return PDF_WRERR;
     }
